+++ release-notes.md
@@ -1,3 +1,4 @@
 # Features
 
 * Established application icon
+* Battery level widget